requirements.txt
numpy<2.0.0
pydub
librosa
pillow
matplotlib
black
